+++ changelog.md
@@ -1,3 +1,7 @@
+## Additions
+- Health bars on the HUD now display the status effects of the target.
+- Added a config option to toggle status effects on the health bar HUD.
+
 ## Bugfixes
 - Damage particles no longer spam the logs.
 - Fixed an issue that caused all nameplates to be permanently visible.